pyproject.toml
[tool.poetry]
name = "hetu"
version = "0.1.0"
description = ""
authors = ["Huayuan Tu <tuhuayuan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
django = "^4.2.6"
django-ninja = "^0.22.2"
django-environ = "^0.11.2"
click = "^8.1.7"
requests = "^2.31.0"
psycopg2 = "^2.9.9"
prometheus-client = "^0.17.1"
python-dateutil = "^2.8.2"
pyjwt = "^2.8.0"
captcha = "^0.5.0"
casbin = "^1.31.2"
casbin-django-orm-adapter = "^1.1.2"
pyyaml = "^6.0.1"
pandas = "^2.1.1"
oss2 = "^2.18.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
url = "http://mirrors.aliyun.com/pypi/simple"
priority = 'default'